data-and-ai-jhb/oct/code/claude-4/expense-chatbot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
azure-ai-projects>=1.0.0
azure-identity==1.19.0
python-multipart==0.0.20
jinja2==3.1.5
python-dotenv==1.0.1
